feat: safety engine, notifications service, route locations, spec 22
Pre-Spec-22 clean slate (Step 0.1):
- internal/safety: driving-safety event detection (harsh driving, idle)
- internal/operations/notifications: notification service + adapters
- 00091_route_locations migration + route geocoder
- ewaybill autogenerate helpers + tests
- handler/view template rework (trip, vehicle, customers portal)
- sqlc regen (v1.27.0) + audit_logs caller fix for NullString param
- docs: product gap analysis, remediation plan, verification notes
- spec 22-command-center-ux (reserved migrations 00092-00094)

+++ internal/handlers/invoices.go
@@ -122,7 +122,7 @@ func (h *InvoiceHandlers) View(w http.ResponseWriter, r *http.Request) {
 	balance, errBal := h.Services.Invoices.GetBalance(r.Context(), domain.InvoiceID(idParam))
 	if errBal != nil {
 		slog.Error("Failed to calculate balance for invoice", "invoice_id", idParam, "error", errBal)
-		balance = 0
+		balance = float64(0)
 	}
 
 	session, _ := h.getUserFromContext(r)
